chore: update SampleDataGenerator with new APIs
